palette/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
pandas==2.2.2
scikit-learn==1.5.2
numpy==2.0.2
requests==2.32.3
streamlit==1.39.0
